requirements.txt
selenium==4.14.0
pytest==7.4.2
pytest-mypy==0.10.3
python-dotenv